Added first section of language files

+++ src/commands/Fun/advice.js
@@ -2,14 +2,14 @@
 const { Random } = require('something-random-on-discord');
 const random = new Random();
 
-module.exports.run = async (bot, message, args, emojis) => {
+module.exports.run = async (bot, message, args, emojis, settings) => {
 	try {
 		const data = await random.getAdvice();
 		message.channel.send(data);
 	} catch (err) {
 		if (bot.config.debug) bot.logger.error(`${err.message} - command: advice.`);
-		message.channel.send({ embed:{ color:15158332, description:`${emojis[0]} An error occured when running this command, please try again or contact support.` } }).then(m => m.delete({ timeout: 5000 }));
-		message.delete();
+		message.error(settings.Language, 'ERROR_MESSAGE').then(m => m.delete({ timeout: 5000 }));
+		if (message.deletable) message.delete();
 	}
 };
