+++ src/commands/Fun/fact.js
@@ -2,20 +2,20 @@
 const fs = require('fs');
 const { MessageEmbed } = require('discord.js');
 
-module.exports.run = async (bot, message, args, emojis) => {
+module.exports.run = async (bot, message, args, emojis, settings) => {
 	// Get the random facts file
 	fs.readFile('./src/assets/json/random-facts.json', (err, data) => {
 		if (err) {
 			if (bot.config.debug) bot.logger.error(`${err.message} - command: fact.`);
-			message.channel.send({ embed:{ color:15158332, description:`${emojis[0]} An error occured when running this command, please try again or contact support.` } }).then(m => m.delete({ timeout: 5000 }));
-			message.delete();
+			message.error(settings.Language, 'ERROR_MESSAGE').then(m => m.delete({ timeout: 5000 }));
+			if (message.deletable) message.delete();
 			return;
 		}
 		const facts = JSON.parse(data);
 		// Retrieve a random fact
 		const num = (Math.floor((Math.random() * facts.facts.length) + 0));
 		const embed = new MessageEmbed()
-			.setTitle('Random Fact:')
+			.setTitle(message.translate(settings.Language, 'FUN/FACT_TITLE'))
 			.setColor('RANDOM')
 			.setDescription(facts.facts[num]);
 		message.channel.send(embed);
